changed userInfo type to NSDictionary*
git-svn-id: svn://orca.physics.unc.edu/Orca@11826 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Hardware/USB/LabJack/LabJackUE9/ORLabJackUE9Model.h
@@ -206,7 +206,7 @@
 - (void) setIsConnected:(BOOL)aFlag;
 - (void) connect;
 - (void) netsocketConnected:(NetSocket*)inNetSocket;
-- (void) netsocket:(NetSocket*)inNetSocket dataAvailable:(unsigned)inAmount;
+- (void) netsocket:(NetSocket*)inNetSocket dataAvailable:(NSUInteger)inAmount;
 - (void) netsocketDisconnected:(NetSocket*)inNetSocket;
 
 #pragma mark ***HW Access
