+++ Source/Objects/Hardware/USB/LakeShore336/ORLakeShore336Model.h
@@ -130,7 +130,7 @@
 
 #pragma mark ***Delegate Methods
 - (void) netsocketConnected:(NetSocket*)inNetSocket;
-- (void) netsocket:(NetSocket*)inNetSocket dataAvailable:(unsigned)inAmount;
+- (void) netsocket:(NetSocket*)inNetSocket dataAvailable:(NSUInteger)inAmount;
 - (void) netsocketDisconnected:(NetSocket*)inNetSocket;
 
 #pragma mark ***Utilities
